meson.bbclass: do not pass native compiler/linker flags via command line
With 0.51.0 version these command line options override what is in the cross file (e.g.
the cross-flags). I could not determine what is the scenario when the native flags are
needed (this would be building a native binary in the context of cross build).
If we find such a scenario we would need to find a way to pass native flags
through some other channel.

(From OE-Core rev: 6239b77c78b4b584274019130b715f21ac9defdf)

Signed-off-by: Alexander Kanavin <alex.kanavin@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>